requirements.txt
Django==4.0.6
psycopg2-binary==2.9.3
django-bootstrap-form==3.4
django-bootstrap-datepicker-plus==5.0.2
uwsgi==2.0.20
invoke==1.7.1
django-anymail[mailgun]==9.0
celery==5.2.7
redis==4.4.2
django-celery-beat==2.4.0
django-filter==22.1

djangorestframework==3.14.0
drf-yasg==1.21.4
djangorestframework-simplejwt==5.2.2
social-auth-app-django==5.0.0
